Refinement to order of lsf run chain submission and dependencies.
Aim here is to minimize the number of state files stored in the workspace at any given time. Submit the potential runs beginning from restarts of an actual run immediately after that run is submitted, rather than waiting for ALL actual runs of that period. Every potential run deletes the state it started from, unless that state will be used by the next actual run (or if its restart came from the spinup run, or its restart was the first year of the future period). In that case, the next actual run will delete its restart, as long as the potential run in question did not complete successfully.